# src/main/java/ru/egot/diplom/workout/dto/user/UserDto.java
package ru.egot.diplom.workout.dto.user;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder(toBuilder = true)
public class UserDto {

		private String username;

		private String password;
}