# src/main/java/ru/egot/diplom/workout/entity/TrainingPlanEntity.java
package ru.egot.diplom.workout.entity;

import jakarta.persistence.*;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.time.DayOfWeek;

@Entity
@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Table(name = "TRAINING_PLAN")
public class TrainingPlanEntity extends BaseEntity {

    @JoinColumn(name = "USER_ID")
    @NotNull
    @ManyToOne
    private User user;

    @JoinColumn(name = "TRAINING_ID")
    @NotNull
    @ManyToOne
    private TrainingEntity training;

    @NotNull
    @Column(name = "DAY_OF_WEEK")
    @Enumerated(EnumType.STRING)
    private DayOfWeek days;

}